refactor: extract FloatingNavbar logic into useFloatingNavbarState hook
