+++ src/app/routes/HomeRoute.tsx
@@ -9,7 +9,6 @@ import { ErrorBoundary } from "@/shared/components/layout/Feedback/ErrorBoundary
 import { Loading } from "@/shared/components/layout/Feedback/Loading";
 import { Section } from "@/shared/components/layout/Section";
 import { SectionHeading } from "@/shared/components/layout/SectionHeading";
-import { SectionNavigation } from "@/shared/components/layout/SectionNavigation";
 import { useSectionScroll } from "@/shared/hooks/useSectionScroll";
 import { getLockedNames } from "@/shared/lib/names/nameFilters";
 import useAppStore from "@/store/appStore";
@@ -64,7 +63,10 @@ export default function HomeRoute() {
 				<div className="flex flex-col items-center justify-center min-h-[100dvh] py-12 md:py-16">
 					<div className="w-full flex flex-col items-center gap-8 md:gap-12">
 						<div>
-							<SectionHeading title="Pick Your Favorites" subtitle="Swipe or click to select names you love." />
+							<SectionHeading
+								title="Pick Your Favorites"
+								subtitle="Swipe or click to select names you love."
+							/>
 						</div>
 						<div className="w-full">
 							<Suspense fallback={<Loading variant="skeleton" height={400} />}>
@@ -94,7 +96,10 @@ export default function HomeRoute() {
 				<div className="flex flex-col items-center justify-center min-h-[100dvh] py-12 md:py-16">
 					<div className="w-full flex flex-col items-center gap-8 md:gap-12">
 						<div>
-							<SectionHeading title="Compare Names" subtitle="Vote on which name you prefer in each matchup." />
+							<SectionHeading
+								title="Compare Names"
+								subtitle="Vote on which name you prefer in each matchup."
+							/>
 						</div>
 						<Suspense fallback={<Loading variant="skeleton" height={400} />}>
 							{tournament.names && tournament.names.length > 0 ? (
