+++ src/app/routes/components/HomeSections.test.tsx
@@ -52,7 +52,7 @@ describe("HomeHeroSection", () => {
 
 		await renderHomeHeroSection({ onStartPicking });
 
-		fireEvent.click(screen.getByText("Narrow It Down"));
+		fireEvent.click(screen.getByText("Get Started"));
 
 		expect(onStartPicking).toHaveBeenCalledTimes(1);
 	});
