+++ src/features/tournament/modes/TournamentFlow.test.tsx
@@ -48,7 +48,7 @@ describe("TournamentFlow responsive behavior", () => {
 		renderWithProviders();
 
 		expect(screen.getByTestId("name-selector")).toBeInTheDocument();
-		expect(screen.queryByRole("button", { name: "Analyze Results" })).not.toBeInTheDocument();
+		expect(screen.queryByRole("button", { name: "See Results" })).not.toBeInTheDocument();
 	});
 
 	it("keeps completion actions mobile-friendly with stacked buttons", () => {
@@ -57,8 +57,8 @@ describe("TournamentFlow responsive behavior", () => {
 
 		renderWithProviders();
 
-		const analyzeButton = screen.getByRole("button", { name: "Analyze Results" });
-		const startButton = screen.getByRole("button", { name: "Start New Tournament" });
+		const analyzeButton = screen.getByRole("button", { name: "See Results" });
+		const startButton = screen.getByRole("button", { name: "Pick Different Names" });
 		const heading = screen.getByRole("heading", {
 			name: "A victor emerges from the eternal tournament",
 		});
