+++ src/shared/components/layout/FloatingNavbar.tsx
@@ -1,25 +1,12 @@
-import {
-	BarChart3,
-	CheckCircle,
-	Layers,
-	LayoutGrid,
-	Lightbulb,
-	Lock,
-	Trophy,
-	User,
-} from "lucide-react";
-import { lazy, Suspense, useCallback, useEffect, useMemo, useRef, useState } from "react";
-import { useLocation, useNavigate } from "react-router-dom";
-import { useAuth } from "@/app/providers/Providers";
+import { lazy, Suspense } from "react";
 import { Loading } from "@/shared/components/layout/Feedback/Loading";
 import { Modal } from "@/shared/components/layout/Modal";
 import {
 	DynamicIslandNav,
 	type DynamicIslandNavItem,
 } from "@/shared/components/ui/dynamic-island-nav";
-import { hapticNavTap, hapticTournamentStart } from "@/shared/lib/browser/haptics";
 import { cn } from "@/shared/lib/utils";
-import useAppStore from "@/store/appStore";
+import { useFloatingNavbarState } from "./useFloatingNavbarState";
 
 function MobileBottomNav({
 	items,
@@ -68,14 +55,6 @@ function MobileBottomNav({
 	);
 }
 
-type NavSection = "pick" | "tournament" | "analysis";
-
-const keyToId: Record<NavSection, string> = {
-	pick: "pick",
-	tournament: "tournament",
-	analysis: "analysis",
-};
-
 const LazyProfileInner = lazy(() =>
 	import("@/shared/components/profile/ProfileInner").then((module) => ({
 		default: module.ProfileInner,
@@ -89,413 +68,33 @@ const LazyNameSuggestion = lazy(() =>
 );
 
 export function FloatingNavbar() {
-	const appStore = useAppStore();
-	const navigate = useNavigate();
-	const location = useLocation();
-	const { login, logout } = useAuth();
-	const { tournament, tournamentActions, user, ui, uiActions } = appStore;
-	const { selectedNames } = tournament;
-	const { isLoggedIn, name: userName, avatarUrl, isAdmin } = user;
-	const { isSwipeMode } = ui;
-	const { setSwipeMode } = uiActions;
-	const [activeSection, setActiveSection] = useState<NavSection>("pick");
-	const [isNavVisible, setIsNavVisible] = useState(true);
-	const [prefersReducedMotion, setPrefersReducedMotion] = useState(false);
-	const [scrollProgress, setScrollProgress] = useState(0);
-	const [pendingScroll, setPendingScroll] = useState<NavSection | null>(null);
-	const [isProfileOpen, setIsProfileOpen] = useState(false);
-	const [isSuggestOpen, setIsSuggestOpen] = useState(false);
-	const profileButtonRef = useRef<HTMLDivElement | null>(null);
-	const suggestButtonRef = useRef<HTMLDivElement | null>(null);
-	const [profileOriginRect, setProfileOriginRect] = useState<{
-		x: number;
-		y: number;
-		width: number;
-		height: number;
-	} | null>(null);
-	const [suggestOriginRect, setSuggestOriginRect] = useState<{
-		x: number;
-		y: number;
-		width: number;
-		height: number;
-	} | null>(null);
-
-	const isHomeRoute = location.pathname === "/";
-	const isAdminRoute = location.pathname === "/admin";
-	const isTournamentRoute = location.pathname === "/tournament";
-	const [isPastHero, setIsPastHero] = useState(false);
-
-	const selectedCount = selectedNames?.length || 0;
-	const isTournamentActive = Boolean(tournament.names);
-	const profileLabel = isLoggedIn ? userName?.split(" ")[0] || "Profile" : "Profile";
-
-	const scrollToSection = useCallback(
-		(key: NavSection) => {
-			const id = keyToId[key];
-			const target = document.getElementById(id);
-			if (!target) {
-				window.scrollTo({
-					top: 0,
-					behavior: prefersReducedMotion ? "auto" : "smooth",
-				});
-				return;
-			}
-
-			target.scrollIntoView({
-				behavior: prefersReducedMotion ? "auto" : "smooth",
-				block: "start",
-			});
-		},
-		[prefersReducedMotion],
-	);
-
-	const handleStartTournament = useCallback(() => {
-		hapticTournamentStart();
-		if (selectedNames && selectedNames.length >= 2) {
-			tournamentActions.setNames(selectedNames);
-			if (isHomeRoute) {
-				scrollToSection("tournament");
-			} else {
-				setPendingScroll("tournament");
-				navigate("/");
-			}
-		}
-	}, [isHomeRoute, navigate, scrollToSection, selectedNames, tournamentActions]);
-
-	const handleNavClick = useCallback(
-		(key: NavSection) => {
-			hapticNavTap();
-			if (!isHomeRoute) {
-				setPendingScroll(key);
-				navigate("/");
-				return;
-			}
-			scrollToSection(key);
-		},
-		[isHomeRoute, navigate, scrollToSection],
-	);
-
-	const handleAdminClick = useCallback(() => {
-		hapticNavTap();
-		if (!isAdminRoute) {
-			navigate("/admin");
-		}
-	}, [isAdminRoute, navigate]);
-
-	const openProfileModal = useCallback(() => {
-		hapticNavTap();
-		const buttonEl = profileButtonRef.current;
-		if (buttonEl) {
-			const rect = buttonEl.getBoundingClientRect();
-			setProfileOriginRect({
-				x: rect.left,
-				y: rect.top,
-				width: rect.width,
-				height: rect.height,
-			});
-		}
-		setIsProfileOpen(true);
-	}, []);
-
-	const openSuggestModal = useCallback(() => {
-		hapticNavTap();
-		const buttonEl = suggestButtonRef.current;
-		if (buttonEl) {
-			const rect = buttonEl.getBoundingClientRect();
-			setSuggestOriginRect({
-				x: rect.left,
-				y: rect.top,
-				width: rect.width,
-				height: rect.height,
-			});
-		}
-		setIsSuggestOpen(true);
-	}, []);
-
-	useEffect(() => {
-		if (!isHomeRoute || !pendingScroll) {
-			return;
-		}
-		scrollToSection(pendingScroll);
-		setPendingScroll(null);
-	}, [isHomeRoute, pendingScroll, scrollToSection]);
-
-	useEffect(() => {
-		if (!isHomeRoute) {
-			return;
-		}
-
-		let rafId: number | null = null;
-		const sections: NavSection[] = ["pick", "tournament", "analysis"];
-
-		const handleScroll = () => {
-			if (rafId) {
-				return;
-			}
-			rafId = requestAnimationFrame(() => {
-				rafId = null;
-				let current: NavSection = "pick";
-				let minDistance = Number.POSITIVE_INFINITY;
-
-				for (const section of sections) {
-					const element = document.getElementById(section);
-					if (!element) {
-						continue;
-					}
-					const rect = element.getBoundingClientRect();
-					const distance = Math.abs(rect.top);
-					if (distance < minDistance && rect.top < window.innerHeight * 0.7) {
-						minDistance = distance;
-						current = section;
-					}
-				}
-				setActiveSection(current);
-
-				const total = document.documentElement.scrollHeight - window.innerHeight;
-				setScrollProgress(
-					total > 0 ? Math.min(100, Math.max(0, (window.scrollY / total) * 100)) : 0,
-				);
-			});
-		};
-
-		window.addEventListener("scroll", handleScroll, { passive: true });
-		handleScroll();
-		return () => {
-			window.removeEventListener("scroll", handleScroll);
-			if (rafId) {
-				cancelAnimationFrame(rafId);
-			}
-		};
-	}, [isHomeRoute]);
-
-	useEffect(() => {
-		if (!isHomeRoute) {
-			setScrollProgress(0);
-		}
-	}, [isHomeRoute]);
-
-	useEffect(() => {
-		const mediaQuery = window.matchMedia("(prefers-reduced-motion: reduce)");
-		const updatePreference = () => setPrefersReducedMotion(mediaQuery.matches);
-		updatePreference();
-		mediaQuery.addEventListener("change", updatePreference);
-		return () => mediaQuery.removeEventListener("change", updatePreference);
-	}, []);
-
-	useEffect(() => {
-		if (!isHomeRoute) {
-			setIsPastHero(true);
-			return;
-		}
-		const check = () => {
-			setIsPastHero(window.scrollY > window.innerHeight * 0.85);
-		};
-		check();
-		window.addEventListener("scroll", check, { passive: true });
-		return () => {
-			window.removeEventListener("scroll", check);
-			setIsPastHero(false);
-		};
-	}, [isHomeRoute]);
-
-	useEffect(() => {
-		let lastScrollY = window.scrollY;
-		let ticking = false;
-		const mobileMediaQuery = window.matchMedia("(max-width: 768px)");
-
-		const onScroll = () => {
-			if (!mobileMediaQuery.matches) {
-				lastScrollY = window.scrollY;
-				return;
-			}
-
-			if (ticking) {
-				return;
-			}
-
-			ticking = true;
-			requestAnimationFrame(() => {
-				const currentScrollY = window.scrollY;
-				const delta = currentScrollY - lastScrollY;
-
-				if (delta > 12) {
-					setIsNavVisible(false);
-				} else if (delta < -12) {
-					setIsNavVisible(true);
-				}
-
-				lastScrollY = currentScrollY;
-				ticking = false;
-			});
-		};
-
-		const onViewportChange = () => {
-			if (!mobileMediaQuery.matches) {
-				setIsNavVisible(true);
-			}
-		};
-
-		window.addEventListener("scroll", onScroll, { passive: true });
-		mobileMediaQuery.addEventListener("change", onViewportChange);
-
-		return () => {
-			window.removeEventListener("scroll", onScroll);
-			mobileMediaQuery.removeEventListener("change", onViewportChange);
-		};
-	}, []);
-
-	const primaryLabel = useMemo(() => {
-		if (isAdminRoute) {
-			return "Admin";
-		}
-		if (!isHomeRoute) {
-			return "Navigation";
-		}
-		if (isTournamentActive) {
-			return "Tournament";
-		}
-		if (selectedCount >= 2) {
-			return `Start (${selectedCount})`;
-		}
-		if (activeSection === "analysis") {
-			return "Analyze";
-		}
-		if (activeSection === "tournament") {
-			return "Bracket";
-		}
-		return "Pick Names";
-	}, [activeSection, isAdminRoute, isHomeRoute, isTournamentActive, selectedCount]);
-
-	const navItems = useMemo((): DynamicIslandNavItem[] => {
-		const items: DynamicIslandNavItem[] = [];
-
-		if (isHomeRoute) {
-			items.push({
-				id: "pick",
-				label: isTournamentActive
-					? "Compare"
-					: selectedCount >= 2
-						? `Vote (${selectedCount})`
-						: "Favorites",
-				level: 1,
-				icon: isTournamentActive ? (
-					<Trophy className="h-4 w-4" />
-				) : selectedCount >= 2 ? (
-					<Trophy className="h-4 w-4" />
-				) : (
-					<CheckCircle className="h-4 w-4" />
-				),
-				isActive: activeSection === "pick" || activeSection === "tournament",
-				isAccent: isTournamentActive || selectedCount >= 2,
-				onClick: () => {
-					if (isTournamentActive) {
-						handleNavClick("tournament");
-					} else if (selectedCount >= 2) {
-						handleStartTournament();
-					} else {
-						handleNavClick("pick");
-					}
-				},
-			});
-
-			items.push({
-				id: "analysis",
-				label: "Results",
-				level: 1,
-				icon: <BarChart3 className="h-4 w-4" />,
-				isActive: activeSection === "analysis",
-				hasBadge: Object.keys(tournament.ratings).length > 0 && activeSection !== "analysis",
-				onClick: () => handleNavClick("analysis"),
-			});
-		}
-
-		items.push({
-			id: "suggest",
-			label: "Suggest",
-			level: isHomeRoute ? 1 : 1,
-			icon: <Lightbulb className="h-4 w-4" />,
-			isActive: isSuggestOpen,
-			onClick: openSuggestModal,
-		});
-
-		if (isAdmin) {
-			items.push({
-				id: "admin",
-				label: "Admin",
-				level: 1,
-				icon: <Lock className="h-4 w-4" />,
-				isActive: isAdminRoute,
-				onClick: handleAdminClick,
-			});
-		}
-
-		items.push({
-			id: "profile",
-			label: profileLabel,
-			level: 1,
-			icon:
-				isLoggedIn && avatarUrl ? (
-					<img
-						src={avatarUrl}
-						alt={profileLabel}
-						className="h-5 w-5 rounded-full border border-foreground/15 object-cover"
-					/>
-				) : (
-					<User
-						className={cn(
-							"h-4 w-4",
-							isLoggedIn && isAdmin && "text-chart-4",
-							isLoggedIn && !isAdmin && "text-primary",
-						)}
-					/>
-				),
-			isActive: isProfileOpen,
-			onClick: openProfileModal,
-		});
-
-		if (isHomeRoute) {
-			items.push({
-				id: "layout-mode",
-				label: isSwipeMode ? "Swipe mode" : "Grid mode",
-				level: 2,
-				icon: isSwipeMode ? <Layers className="h-4 w-4" /> : <LayoutGrid className="h-4 w-4" />,
-				ariaLabel: isSwipeMode ? "Swipe mode active" : "Grid mode active",
-				ariaPressed: isSwipeMode,
-				onClick: () => {
-					hapticNavTap();
-					setSwipeMode(!isSwipeMode);
-				},
-			});
-		}
-
-		return items;
-	}, [
+	const {
+		isTournamentRoute,
+		shouldShow,
+		profileButtonRef,
+		suggestButtonRef,
+		navItems,
+		primaryLabel,
 		activeSection,
-		avatarUrl,
-		handleAdminClick,
-		handleNavClick,
-		handleStartTournament,
-		isAdmin,
+		selectedCount,
+		isTournamentActive,
 		isAdminRoute,
 		isHomeRoute,
-		isLoggedIn,
+		scrollProgress,
+		prefersReducedMotion,
 		isProfileOpen,
+		setIsProfileOpen,
+		profileOriginRect,
 		isSuggestOpen,
-		isSwipeMode,
-		isTournamentActive,
-		openProfileModal,
-		openSuggestModal,
-		profileLabel,
-		selectedCount,
-		setSwipeMode,
-		tournament.ratings,
-	]);
+		setIsSuggestOpen,
+		suggestOriginRect,
+		login,
+		logout,
+	} = useFloatingNavbarState();
 
 	if (isTournamentRoute) {
 		return null;
 	}
-
-	const shouldShow = isNavVisible && (!isHomeRoute || isPastHero);
 
 	return (
 		<>
